05-Projects/00-terminal-prototype/pyproject.toml
[project]
name = "ai-logging-agent-terminal"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "openai-agents[litellm]>=0.0.19",
    "python-dotenv",
]